requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# Redis
redis==5.0.1
aioredis==2.0.1

# AI APIs
google-generativeai==0.3.1
openai==1.3.7
groq==0.4.1

# WhatsApp (Twilio)
twilio==8.10.3

# Telegram
python-telegram-bot==20.7

# Utils
requests==2.31.0
httpx==0.25.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging
python-json-logger==2.0.7